usb3503: add optional ext-regulator property
On some developer boards where the USB3503 is used, additional blocks
are connected to it. E.g. on the Hardkernel ODROID-U2 a LAN97xx
(ethernet with USB hub combo) is connected to it. The LAN97xx needs
its regulator enabled, but in particular this regulator also needs
to be switched off again on shutdown, otherwise it the LAN97xx is no
longer detected after a reboot.

This patch adds a new DT property 'ext-supply' so that one can
specify which regulator has to be enabled on probe / disabled on
shutdown.